requirements.txt
requests
Pillow
python-dotenv
RPi.GPIO
spidev
pytest